requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
psycopg2-binary
langchain-text-splitters
tiktoken